"Refactor and separate concerns for improved maintainability"
In this commit, the codebase has been refactored to achieve improved maintainability and easier modification in the future.

The HTTP request/responses, processing of Location and CurrentConditions, and introducing the ability to read API keys from properties file have been separated out into their individual classes ("AccuweatherClient", "AccuweatherService", and "PropertiesReaderUtils").

This change enhances the separation of concerns, making each class responsible for specific tasks. The "HttpExample" class is also simplified, making it easier to understand the overall flow of the application. Deletions of unnecessary code within certain classes also helped to reduce clutter and improve readability of the code.

Furthermore, a properties file has been introduced to store API keys, which makes it easier for the keys to be changed and managed. This avoids having the keys hard-coded in the application.